Add callout lines for region and country-wide averages
This also introduces a new function to generate bucket intervals to
replace D3's ticks(), which generates human-readable marks for a graph,
aimted at generating pleasing numbers that roughly divide the range in
the requested number of segments. However, because these numbers are
meant to be human-readable, they don't fully include the entire range...
if valur range is [0, 8351], for instance, dividing that into 100
segments may result in a list like [0, 100, 200, ..., 8300], leaving
the max value of 8351 to literally be off the chart.

The new method always splits the value range into exactly N buckets
of equal size, making sure that extreme values are always included in
the graph.