Add legend for histogram
Legend is rendered dynamically based on which callouts have values. This
involved refactoring the way callouts were passed to the object to keep
their attributes consistent (Associating name, color and value in one
object), which in turn required refactoring how the callout values were
generated.

The legend labels do not give the cluster name as the front-end doesn't
know what cluster the soum belongs to (A service to look that up would
need to be created). The soum name is shown, though this requires
waiting for both the comparison request and the base soum request to
complete. This isn't a significant barrier, though, as the rest of the
page depends on the soum data and a comparative graph doesn't have much
value without any data around it.